# backend/CorporateCashFlow.Entity/DTOs/ErrorResponseDto.cs
namespace CorporateCashFlow.Entity.DTOs;

public class ErrorResponseDto
{
    public string? Type { get; set; }
    public string? Title { get; set; }
    public int Status { get; set; }
    public string? Detail { get; set; }
    public string? Instance { get; set; }
    public Dictionary<string, string[]>? Errors { get; set; }
}