# backend/CorporateCashFlow.Entity/DTOs/LoginRequestDto.cs
using System.ComponentModel.DataAnnotations;

namespace CorporateCashFlow.Entity.DTOs;

public class LoginRequestDto
{
    [Required]
    [EmailAddress]
    public string Email { get; set; } = string.Empty;

    [Required]
    [MinLength(8)]
    public string Password { get; set; } = string.Empty;
}