# backend/CorporateCashFlow.Entity/DTOs/LoginResponseDto.cs
namespace CorporateCashFlow.Entity.DTOs;

public class LoginResponseDto
{
    public string Token { get; set; } = string.Empty;
    public string RefreshToken { get; set; } = string.Empty;
    public DateTimeOffset ExpiresAt { get; set; }
}