# backend/CorporateCashFlow.Entity/DTOs/TokenRefreshResponseDto.cs
namespace CorporateCashFlow.Entity.DTOs;

public class TokenRefreshResponseDto
{
    public string AccessToken { get; set; } = string.Empty;
    public string RefreshToken { get; set; } = string.Empty;
    public DateTimeOffset ExpiresAt { get; set; }
}